+++ src/ts/Managers/CampaignManager.ts
@@ -11,7 +11,7 @@ import { MetaDataManager } from 'Managers/MetaDataManager';
 import { JsonParser } from 'Utilities/JsonParser';
 import { DiagnosticError } from 'Errors/DiagnosticError';
 import { StorageType } from 'Native/Api/Storage';
-import { HtmlCampaign } from 'Models/HtmlCampaign';
+import { MRAIDCampaign } from 'Models/MRAIDCampaign';
 import { PerformanceCampaign } from 'Models/PerformanceCampaign';
 import { AssetManager } from 'Managers/AssetManager';
 import { WebViewError } from 'Errors/WebViewError';
@@ -48,7 +48,7 @@ export class CampaignManager {
 
     public onPerformanceCampaign: Observable1<PerformanceCampaign> = new Observable1();
     public onVastCampaign: Observable1<VastCampaign> = new Observable1();
-    public onThirdPartyCampaign: Observable1<HtmlCampaign> = new Observable1();
+    public onMRAIDCampaign: Observable1<MRAIDCampaign> = new Observable1();
     public onNoFill: Observable1<number> = new Observable1();
     public onError: Observable1<WebViewError> = new Observable1();
 
@@ -108,6 +108,7 @@ export class CampaignManager {
         if(json.gamerId) {
             this.storeGamerId(json.gamerId);
         }
+
         if('campaign' in json) {
             return this.parsePerformanceCampaign(json);
         } else if('vast' in json) {
@@ -119,56 +120,13 @@ export class CampaignManager {
 
     private parsePerformanceCampaign(json: any): Promise<void> {
         this._nativeBridge.Sdk.logInfo('Unity Ads server returned game advertisement for AB Group ' + json.abGroup);
-        const htmlCampaign = this.parseHtmlCampaign(json);
-        if(htmlCampaign) {
-            return this._assetManager.setup(htmlCampaign).then(() => this.onThirdPartyCampaign.trigger(htmlCampaign));
+        if(json.campaign && json.campaign.mraidUrl) {
+            const campaign = new MRAIDCampaign(json.campaign, json.gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : json.abGroup, json.campaign.mraidUrl);
+            return this._assetManager.setup(campaign).then(() => this.onMRAIDCampaign.trigger(campaign));
         } else {
             const campaign = new PerformanceCampaign(json.campaign, json.gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : json.abGroup);
             return this._assetManager.setup(campaign).then(() => this.onPerformanceCampaign.trigger(campaign));
         }
-    }
-
-    private parseHtmlCampaign(json: any): HtmlCampaign | undefined {
-        const campaign = new PerformanceCampaign(json.campaign, json.gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : json.abGroup);
-        let resource: string | undefined;
-        switch(campaign.getId()) {
-            // Game of War iOS
-            case '583dfda0d933a3630a53249c':
-            case '583dfd52abb1feee0909882b':
-            case '583dfd45669a903e086e38d2':
-            case '583dfd4c9ceadb4708b021de':
-                resource = 'https://static.applifier.com/playables/SG_ios/index_ios.html';
-                break;
-
-            // Game of War Android
-            case '583dfca5a93bfa6700d8c6f3':
-            case '583dfcb54622865a0a246bdf':
-                resource = 'https://static.applifier.com/playables/SG_android/index_android.html';
-                break;
-
-            // Mobile Strike iOS
-            case '583dfb9a5b79df3f0a274f0b':
-            case '583dfe483fe2166c0ac9e6fb':
-            case '583dfba09bfc2a2d0a9a0b1c':
-            case '583dfba69d308fe203d7d740':
-                resource = 'https://static.applifier.com/playables/SMA_ios/index_ios.html';
-                break;
-
-            // Mobile Strike Android
-            case '583dfc532e4d9b5008c934d1':
-            case '583dfc667f448e630ac6a4bc':
-                resource = 'https://static.applifier.com/playables/SMA_android/index_android.html';
-                break;
-
-            default:
-                break;
-        }
-
-        const abGroup = campaign.getAbGroup();
-        if(resource && abGroup !== 6 && abGroup !== 7) {
-            return new HtmlCampaign(json.campaign, json.gamerId, CampaignManager.AbGroup ? CampaignManager.AbGroup : json.abGroup, resource);
-        }
-        return undefined;
     }
 
     private parseVastCampaign(json: any): Promise<void> {
@@ -211,7 +169,7 @@ export class CampaignManager {
                 this.onError.trigger(videoUrlError);
                 return;
             }
-            this._assetManager.setup(campaign).then(() => this.onVastCampaign.trigger(campaign));
+            return this._assetManager.setup(campaign).then(() => this.onVastCampaign.trigger(campaign));
         }).catch((error) => {
             this.onError.trigger(error);
         });
@@ -250,7 +208,8 @@ export class CampaignManager {
             screenWidth: this.getParameter('screenWidth', this._deviceInfo.getScreenWidth(), 'number'),
             screenHeight: this.getParameter('screenHeight', this._deviceInfo.getScreenHeight(), 'number'),
             sdkVersion: this.getParameter('sdkVersion', this._clientInfo.getSdkVersion(), 'number'),
-            screenSize: this.getParameter('screenSize', this._deviceInfo.getScreenLayout(), 'number')
+            screenSize: this.getParameter('screenSize', this._deviceInfo.getScreenLayout(), 'number'),
+            stores: this.getParameter('stores', this._deviceInfo.getStores(), 'string')
         });
 
         if(this._clientInfo.getPlatform() === Platform.IOS) {