+++ src/ts/WebView.ts
@@ -75,9 +75,8 @@ export class WebView {
         return this._nativeBridge.Sdk.loadComplete().then((data) => {
             this._deviceInfo = new DeviceInfo(this._nativeBridge);
             this._wakeUpManager = new WakeUpManager(this._nativeBridge);
-            this._cache = new Cache(this._nativeBridge, this._wakeUpManager);
-            this._cache.cleanCache();
             this._request = new Request(this._nativeBridge, this._wakeUpManager);
+            this._cache = new Cache(this._nativeBridge, this._wakeUpManager, this._request);
             this._resolve = new Resolve(this._nativeBridge);
             this._clientInfo = new ClientInfo(this._nativeBridge.getPlatform(), data);
             this._eventManager = new EventManager(this._nativeBridge, this._request);
@@ -85,6 +84,8 @@ export class WebView {
             HttpKafka.setClientInfo(this._clientInfo);
 
             return this._deviceInfo.fetch();
+        }).then(() => {
+            return this._cache.cleanCache();
         }).then(() => {
             if(this._clientInfo.getPlatform() === Platform.ANDROID) {
                 document.body.classList.add('android');
@@ -122,12 +123,13 @@ export class WebView {
         }).then(() => {
             const defaultPlacement = this._configuration.getDefaultPlacement();
             this._nativeBridge.Placement.setDefaultPlacement(defaultPlacement.getId());
-            this.setPlacementStates(PlacementState.NOT_AVAILABLE);
+            this.setPlacementStates(PlacementState.WAITING);
+
             this._campaignManager = new CampaignManager(this._nativeBridge, new AssetManager(this._cache, this._configuration.getCacheMode()), this._request, this._clientInfo, this._deviceInfo, new VastParser());
             this._campaignManager.onPerformanceCampaign.subscribe(campaign => this.onCampaign(campaign));
             this._campaignManager.onVastCampaign.subscribe(campaign => this.onCampaign(campaign));
-            this._campaignManager.onThirdPartyCampaign.subscribe(campaign => this.onCampaign(campaign));
-            this._campaignManager.onNoFill.subscribe(() => this.onNoFill());
+            this._campaignManager.onMRAIDCampaign.subscribe(campaign => this.onCampaign(campaign));
+            this._campaignManager.onNoFill.subscribe(retryLimit => this.onNoFill());
             this._campaignManager.onError.subscribe(error => this.onCampaignError(error));
             return this._campaignManager.request();
         }).then(() => {
@@ -250,9 +252,14 @@ export class WebView {
         for(const placementId in placements) {
             if(placements.hasOwnProperty(placementId)) {
                 const placement: Placement = placements[placementId];
-                this._nativeBridge.Placement.setPlacementState(placement.getId(), placementState);
+                const oldState = placement.getState();
+                this._nativeBridge.Placement.setPlacementState(placementId, placementState);
+                if(oldState !== placementState) {
+                    this._nativeBridge.Listener.sendPlacementStateChangedEvent(placementId, PlacementState[oldState], PlacementState[placementState]);
+                    placement.setState(placementState);
+                }
                 if(placementState === PlacementState.READY) {
-                    this._nativeBridge.Listener.sendReadyEvent(placement.getId());
+                    this._nativeBridge.Listener.sendReadyEvent(placementId);
                 }
             }
         }